+++ app/Livewire/Layout/Navigation.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Livewire\Layout;
+
+use App\Livewire\Actions\Logout;
+use Livewire\Component;
+
+class Navigation extends Component
+{
+    /**
+     * Log the current user out of the application.
+     */
+    public function logout(Logout $logout): void
+    {
+        $logout();
+
+        $this->redirect('/', navigate: true);
+    }
+
+    public function render()
+    {
+        return view('livewire.layout.navigation');
+    }
+}
