+++ app/Livewire/Pages/Auth/ForgotPassword.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Livewire\Pages\Auth;
+
+use Illuminate\Support\Facades\Password;
+use Livewire\Attributes\Layout;
+use Livewire\Component;
+
+#[Layout('layouts.guest')] class ForgotPassword extends Component
+{
+    public string $email = '';
+
+    /**
+     * Send a password reset link to the provided email address.
+     */
+    public function sendPasswordResetLink(): void
+    {
+        $this->validate([
+            'email' => ['required', 'string', 'email'],
+        ]);
+
+        $status = Password::sendResetLink(
+            $this->only('email')
+        );
+
+        if ($status !== Password::RESET_LINK_SENT) {
+            $this->addError('email', __($status));
+
+            return;
+        }
+
+        $this->reset('email');
+
+        session()->flash('status', __($status));
+    }
+
+    public function render()
+    {
+        return view('livewire.pages.auth.forgot-password');
+    }
+}
