+++ app/Livewire/Pages/Auth/Login.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\Livewire\Pages\Auth;
+
+use App\Livewire\Forms\LoginForm;
+use Illuminate\Support\Facades\Session;
+use Livewire\Attributes\Layout;
+use Livewire\Component;
+
+#[Layout('layouts.guest')] class Login extends Component
+{
+    public LoginForm $form;
+
+    /**
+     * Handle an incoming authentication request.
+     */
+    public function login(): void
+    {
+        $this->validate();
+
+        $this->form->authenticate();
+
+        Session::regenerate();
+
+        $this->redirectIntended(default: route('dashboard', absolute: false), navigate: true);
+    }
+
+    public function render()
+    {
+        return view('livewire.pages.auth.login');
+    }
+}
