+++ app/Livewire/Pages/Auth/Register.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace App\Livewire\Pages\Auth;
+
+use App\Models\User;
+use Illuminate\Auth\Events\Registered;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\Rules\Password;
+use Livewire\Attributes\Layout;
+use Livewire\Component;
+
+#[Layout('layouts.guest')] class Register extends Component
+{
+    public string $name;
+
+    public string $email;
+
+    public string $password;
+
+    public string $password_confirmation;
+
+    /**
+     * Handle an incoming registration request.
+     */
+    public function register(): void
+    {
+        $validated = $this->validate([
+            'name' => ['required', 'string', 'max:255'],
+            'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:'.User::class],
+            'password' => ['required', 'string', 'confirmed', Password::defaults()],
+        ]);
+
+        $validated['password'] = Hash::make($validated['password']);
+
+        event(new Registered($user = User::create($validated)));
+
+        Auth::login($user);
+
+        $this->redirect(route('dashboard', absolute: false), navigate: true);
+    }
+
+    public function render()
+    {
+        return view('livewire.pages.auth.register');
+    }
+}
