+++ routes/web.php
@@ -1,7 +1,6 @@
 <?php
 
 use App\Http\Controllers\LabelController;
-use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
 Route::view('/', 'welcome');
