+++ tests/Feature/Auth/PasswordConfirmationTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use App\Models\User;
-use Livewire\Volt\Volt;
+use Livewire\Livewire;
 
 test('confirm password screen can be rendered', function () {
     $user = User::factory()->create();
@@ -9,7 +9,7 @@ test('confirm password screen can be rendered', function () {
     $response = $this->actingAs($user)->get('/confirm-password');
 
     $response
-        ->assertSeeVolt('pages.auth.confirm-password')
+        ->assertSeeLivewire('pages.auth.confirm-password')
         ->assertStatus(200);
 });
 
@@ -18,7 +18,7 @@ test('password can be confirmed', function () {
 
     $this->actingAs($user);
 
-    $component = Volt::test('pages.auth.confirm-password')
+    $component = Livewire::test('pages.auth.confirm-password')
         ->set('password', 'password');
 
     $component->call('confirmPassword');
@@ -33,7 +33,7 @@ test('password is not confirmed with invalid password', function () {
 
     $this->actingAs($user);
 
-    $component = Volt::test('pages.auth.confirm-password')
+    $component = Livewire::test('pages.auth.confirm-password')
         ->set('password', 'wrong-password');
 
     $component->call('confirmPassword');
