+++ tests/Feature/Auth/PasswordResetTest.php
@@ -5,13 +5,13 @@ namespace Tests\Feature\Auth;
 use App\Models\User;
 use Illuminate\Auth\Notifications\ResetPassword;
 use Illuminate\Support\Facades\Notification;
-use Livewire\Volt\Volt;
+use Livewire\Livewire;
 
 test('reset password link screen can be rendered', function () {
     $response = $this->get('/forgot-password');
 
     $response
-        ->assertSeeVolt('pages.auth.forgot-password')
+        ->assertSeeLivewire('pages.auth.forgot-password')
         ->assertStatus(200);
 });
 
@@ -20,7 +20,7 @@ test('reset password link can be requested', function () {
 
     $user = User::factory()->create();
 
-    Volt::test('pages.auth.forgot-password')
+    Livewire::test('pages.auth.forgot-password')
         ->set('email', $user->email)
         ->call('sendPasswordResetLink');
 
@@ -32,7 +32,7 @@ test('reset password screen can be rendered', function () {
 
     $user = User::factory()->create();
 
-    Volt::test('pages.auth.forgot-password')
+    Livewire::test('pages.auth.forgot-password')
         ->set('email', $user->email)
         ->call('sendPasswordResetLink');
 
@@ -40,7 +40,7 @@ test('reset password screen can be rendered', function () {
         $response = $this->get('/reset-password/'.$notification->token);
 
         $response
-            ->assertSeeVolt('pages.auth.reset-password')
+            ->assertSeeLivewire('pages.auth.reset-password')
             ->assertStatus(200);
 
         return true;
@@ -52,12 +52,12 @@ test('password can be reset with valid token', function () {
 
     $user = User::factory()->create();
 
-    Volt::test('pages.auth.forgot-password')
+    Livewire::test('pages.auth.forgot-password')
         ->set('email', $user->email)
         ->call('sendPasswordResetLink');
 
     Notification::assertSentTo($user, ResetPassword::class, function ($notification) use ($user) {
-        $component = Volt::test('pages.auth.reset-password', ['token' => $notification->token])
+        $component = Livewire::test('pages.auth.reset-password', ['token' => $notification->token])
             ->set('email', $user->email)
             ->set('password', 'password')
             ->set('password_confirmation', 'password');
