+++ tests/Feature/Auth/PasswordUpdateTest.php
@@ -4,14 +4,14 @@ namespace Tests\Feature\Auth;
 
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
-use Livewire\Volt\Volt;
+use Livewire\Livewire;
 
 test('password can be updated', function () {
     $user = User::factory()->create();
 
     $this->actingAs($user);
 
-    $component = Volt::test('profile.update-password-form')
+    $component = Livewire::test('profile.update-password-form')
         ->set('current_password', 'password')
         ->set('password', 'new-password')
         ->set('password_confirmation', 'new-password')
@@ -29,7 +29,7 @@ test('correct password must be provided to update password', function () {
 
     $this->actingAs($user);
 
-    $component = Volt::test('profile.update-password-form')
+    $component = Livewire::test('profile.update-password-form')
         ->set('current_password', 'wrong-password')
         ->set('password', 'new-password')
         ->set('password_confirmation', 'new-password')
