+++ tests/Feature/Auth/RegistrationTest.php
@@ -2,18 +2,18 @@
 
 namespace Tests\Feature\Auth;
 
-use Livewire\Volt\Volt;
+use Livewire\Livewire;
 
 test('registration screen can be rendered', function () {
     $response = $this->get('/register');
 
     $response
         ->assertOk()
-        ->assertSeeVolt('pages.auth.register');
+        ->assertSeeLivewire('pages.auth.register');
 });
 
 test('new users can register', function () {
-    $component = Volt::test('pages.auth.register')
+    $component = Livewire::test('pages.auth.register')
         ->set('name', 'Test User')
         ->set('email', 'test@example.com')
         ->set('password', 'password')
